refactor: 메서드 구조 개선 및 패키지 이동
