+++ backend/src/main/java/com/carffeine/carffeine/car/infrastructure/repository/CarQueryRepository.java
@@ -1,10 +1,11 @@
 package com.carffeine.carffeine.car.infrastructure.repository;
 
-import com.carffeine.carffeine.car.controller.dto.CarResponse;
-import com.carffeine.carffeine.car.controller.dto.CarsResponse;
+import com.carffeine.carffeine.car.infrastructure.dto.CarResponse;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 import static com.carffeine.carffeine.car.domain.QCar.car;
 import static com.querydsl.core.types.Projections.constructor;
@@ -15,16 +16,14 @@ public class CarQueryRepository {
 
     private final JPAQueryFactory jpaQueryFactory;
 
-    public CarsResponse findAllCars() {
-        return new CarsResponse(
-                jpaQueryFactory.select(
-                                constructor(CarResponse.class,
-                                        car.id,
-                                        car.name,
-                                        car.vintage
-                                )
-                        ).from(car)
-                        .fetch()
-        );
+    public List<CarResponse> findAllCars() {
+        return jpaQueryFactory.select(
+                        constructor(CarResponse.class,
+                                car.id,
+                                car.name,
+                                car.vintage
+                        )
+                ).from(car)
+                .fetch();
     }
 }
