--- backend/src/main/java/com/carffeine/carffeine/car/infrastructure/repository/FilterResponse.java
@@ -1,6 +1,0 @@
-package com.carffeine.carffeine.car.infrastructure.repository;
-
-import com.carffeine.carffeine.filter.domain.FilterType;
-
-public record FilterResponse(String name, FilterType filterType) {
-}
