+++ backend/src/main/java/com/carffeine/carffeine/car/service/CarQueryService.java
@@ -1,6 +1,6 @@
 package com.carffeine.carffeine.car.service;
 
-import com.carffeine.carffeine.car.controller.dto.CarsResponse;
+import com.carffeine.carffeine.car.infrastructure.dto.CarsResponse;
 import com.carffeine.carffeine.car.infrastructure.repository.CarQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,6 +14,6 @@ public class CarQueryService {
     private final CarQueryRepository carQueryRepository;
 
     public CarsResponse findAllCars() {
-        return carQueryRepository.findAllCars();
+        return new CarsResponse(carQueryRepository.findAllCars());
     }
 }
