+++ backend/src/main/java/com/carffeine/carffeine/filter/controller/dto/FiltersResponse.java
@@ -1,6 +1,5 @@
 package com.carffeine.carffeine.filter.controller.dto;
 
-import com.carffeine.carffeine.car.infrastructure.repository.FilterResponse;
 import com.carffeine.carffeine.filter.domain.Filter;
 import com.carffeine.carffeine.member.domain.MemberFilter;
 
@@ -14,10 +13,6 @@ public record FiltersResponse(
 
     public static FiltersResponse from(List<Filter> filters) {
         return getResponse(filters);
-    }
-
-    public static FiltersResponse fromFilterResponse(List<FilterResponse> filters) {
-        return getFiltersResponseFromFilterResponse(filters);
     }
 
     private static FiltersResponse getResponse(final List<Filter> filters) {
@@ -41,14 +36,6 @@ public record FiltersResponse(
                 capacities,
                 connectorTypes
         );
-    }
-
-    private static FiltersResponse getFiltersResponseFromFilterResponse(List<FilterResponse> filtersResponse) {
-        List<Filter> filters = filtersResponse.stream()
-                .map(it -> Filter.of(it.name(), it.filterType().getName()))
-                .toList();
-
-        return getResponse(filters);
     }
 
     public static FiltersResponse fromMemberFilters(List<MemberFilter> memberFilters) {
