+++ backend/src/main/java/com/carffeine/carffeine/member/controller/MemberController.java
@@ -1,8 +1,8 @@
 package com.carffeine.carffeine.member.controller;
 
 import com.carffeine.carffeine.auth.controller.support.AuthMember;
-import com.carffeine.carffeine.car.controller.dto.CarResponse;
 import com.carffeine.carffeine.car.domain.Car;
+import com.carffeine.carffeine.car.infrastructure.dto.CarResponse;
 import com.carffeine.carffeine.car.service.dto.CarRequest;
 import com.carffeine.carffeine.filter.controller.dto.FiltersResponse;
 import com.carffeine.carffeine.filter.domain.Filter;
