+++ backend/src/test/java/com/carffeine/carffeine/car/controller/CarControllerTest.java
@@ -1,8 +1,8 @@
 package com.carffeine.carffeine.car.controller;
 
-import com.carffeine.carffeine.car.controller.dto.CarResponse;
-import com.carffeine.carffeine.car.controller.dto.CarsResponse;
 import com.carffeine.carffeine.car.domain.Car;
+import com.carffeine.carffeine.car.infrastructure.dto.CarResponse;
+import com.carffeine.carffeine.car.infrastructure.dto.CarsResponse;
 import com.carffeine.carffeine.car.service.dto.CarRequest;
 import com.carffeine.carffeine.car.service.dto.CarsRequest;
 import com.carffeine.carffeine.filter.controller.dto.FiltersResponse;
