+++ backend/src/test/java/com/carffeine/carffeine/car/infrastructure/repository/CarQueryRepositoryTest.java
@@ -1,8 +1,8 @@
 package com.carffeine.carffeine.car.infrastructure.repository;
 
-import com.carffeine.carffeine.car.controller.dto.CarsResponse;
 import com.carffeine.carffeine.car.domain.Car;
 import com.carffeine.carffeine.car.domain.CarRepository;
+import com.carffeine.carffeine.car.infrastructure.dto.CarResponse;
 import com.carffeine.carffeine.config.QuerydslConfig;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
@@ -10,6 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
+
+import java.util.List;
 
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
@@ -32,12 +34,12 @@ class CarQueryRepositoryTest {
         Car car2 = carRepository.save(Car.from("아이오닉6", "2023"));
 
         // when
-        CarsResponse result = carQueryRepository.findAllCars();
+        List<CarResponse> result = carQueryRepository.findAllCars();
 
         // then
         assertSoftly(softly -> {
-            softly.assertThat(result.cars().get(0).name()).isEqualTo(car1.getName());
-            softly.assertThat(result.cars().get(1).name()).isEqualTo(car2.getName());
+            softly.assertThat(result.get(0).name()).isEqualTo(car1.getName());
+            softly.assertThat(result.get(1).name()).isEqualTo(car2.getName());
         });
     }
 }
